packages/workers/deletion/pyproject.toml
[tool.poetry]
name = "deletion-worker"
version = "0.1.0"
description = "Deletion worker for InsightHub"
authors = ["InsightHub Team"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
psycopg2-binary = "^2.9.0"
qdrant-client = "^1.7.0"
minio = "^7.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.6.0"
types-requests = "^2.31.0"
testcontainers = {extras = ["postgres", "rabbitmq", "qdrant", "minio"], version = "^4.0.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
src_paths = ["src", "tests"]
known_first_party = ["src"]

# Mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = false  # Disable unreachable code warnings for now
strict_equality = true
pretty = true
show_error_codes = true
show_error_context = true
files = ["src"]

[[tool.mypy.overrides]]
module = [
    "testcontainers.*",
    "shared.*",
    "pika.*",
    "psycopg2.*",
    "qdrant_client.*",
    "minio.*",
    "packages.workers.deletion.src.main",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
]
disable_error_code = ["no-untyped-def", "var-annotated", "unused-ignore", "import-untyped", "arg-type", "assignment", "return-value", "union-attr", "attr-defined", "dict-item", "comparison-overlap"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]

ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["ARG001", "ARG002"]  # Allow unused arguments in tests
"tests/conftest.py" = ["E402"]  # Allow imports after environment setup

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
pythonpath = ["tests/dummy_shared", "src"]


# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false